Registration now sending info to database

+++ HSA_App/HSA_App/RegistrationPage.xaml.cs
@@ -20,14 +20,14 @@ namespace HSA_App
 			2.) Ensure the string has the correct length to represent an account number
 
 		*/
-		public int checkAccount(String actNum)
+		public Int64 checkAccount(String actNum)
 		{
-			int number;
+			Int64 number;
 
 			//Checks the Account for all numbers (Attempt to convert)
 			try
 			{
-				number = Convert.ToInt32(actNum);
+				number = Convert.ToInt64(actNum);
 			}
 			catch (Exception ex)
 			{
@@ -194,8 +194,10 @@ namespace HSA_App
 				person.FirstName = fname.Text;
 				person.LastName = lname.Text;
 				person.UserName = username.Text;
-				person.AccountNumber = Convert.ToInt32(accountNum.Text);
-				person.HashedPassword = password1.Text;
+                person.AccountNumber = Convert.ToInt64(accountNum.Text);
+                //Debug.WriteLine("\n\n\n\n\nThere was Nour account number!  " + person.AccountNumber);
+
+                person.HashedPassword = password1.Text;
 
 				//person.Birthday = Convert.ToDateTime(dobLabel.Text);
 				var sv = new WebService();
